--- next-app/src/app/(auth)/layout.tsx
@@ -1,5 +1,0 @@
-const AuthLayout = ({ children }:any) => {
-    return <div>{children}</div>;
-};
-
-export default AuthLayout;
