--- next-app/src/app/(main)/layout.tsx
@@ -1,12 +1,0 @@
-import Header from '@/components/Header';
-
-const MainLayout = ({ children }:any) => {
-    return (
-        <div>
-            <Header />
-            <main className='pt-[70px]'>{children}</main>
-        </div>
-    );
-};
-
-export default MainLayout;
