+++ next-app/src/app/top/page.tsx
@@ -1,8 +1,7 @@
-const Top = () => {
+export default function Page() {
     return(
         <div>
             <h1>美味しい</h1>
         </div>
     )
 }
-export default Top;
